Moved MainViewModel to presentation module
