+++ presentation/src/commonMain/kotlin/dev/fobo66/valiutchik/presentation/MainScreenState.kt
@@ -14,15 +14,22 @@
  *    limitations under the License.
  */
 
-package fobo66.exchangecourcesbelarus.ui
+package dev.fobo66.valiutchik.presentation
 
-const val TAG_RATES = "Rates"
-const val TAG_NO_RATES = "No rates"
-const val TAG_RATE_VALUE = "Currency rate value"
-const val TAG_PREFERENCES = "Preferences"
-const val TAG_SNACKBAR = "Snackbar"
-const val TAG_TITLE = "Title"
-const val TAG_DEFAULT_CITY = "Default city"
-const val TAG_UPDATE_INTERVAL = "Update interval"
-const val TAG_LICENSES = "Licenses"
-const val TAG_SLIDER = "Slider"
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.Stable
+
+@Stable
+sealed class MainScreenState {
+    @Immutable
+    object Initial : MainScreenState()
+
+    @Immutable
+    object Loading : MainScreenState()
+
+    @Immutable
+    object LoadedRates : MainScreenState()
+
+    @Immutable
+    object Error : MainScreenState()
+}
